+++ lib/rpmsg_lite/porting/platform/imxrt500_fusionf1/rpmsg_platform.c
@@ -14,7 +14,7 @@
 #include <xtensa/config/core.h>
 
 #ifdef SDK_OS_BAREMETAL
-#include <xtensa\xtruntime.h>
+#include <xtensa/xtruntime.h>
 #include <xtensa/tie/xt_interrupt.h>
 #else
 #include <xtensa/xos.h>
@@ -227,9 +227,9 @@ void platform_cache_disable(void)
  * Dummy implementation
  *
  */
-uint32_t platform_vatopa(void *addr)
+uintptr_t platform_vatopa(void *addr)
 {
-    return ((uint32_t)(char *)addr);
+    return ((uintptr_t)(char *)addr);
 }
 
 /**
@@ -238,7 +238,7 @@ uint32_t platform_vatopa(void *addr)
  * Dummy implementation
  *
  */
-void *platform_patova(uint32_t addr)
+void *platform_patova(uintptr_t addr)
 {
     return ((void *)(char *)addr);
 }
